requirements.txt
absl-py==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bayesiancoresets==0.9
bleach==1.5.0
certifi==2019.6.16
cycler==0.10.0
decorator==4.4.0
edward==1.3.5
gpflow==1.4.1
h5py==2.9.0
html5lib==0.9999999
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
more-itertools==7.1.0
multipledispatch==0.6.0
numpy==1.16.4
packaging==19.0
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.0
pytest==4.6.3
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
